+++ cmd/ccd/ast/sql.go
@@ -102,8 +102,8 @@ func (s *SelectStmt) sql(db *dbx.DB, a, b *strings.Builder) error {
 	var projection string
 	switch s.AttrList.(*SelectAttrList).Attr {
 	case "*":
-		// projection = "a.id, coalesce(a.author, '') as author, coalesce(a.title, '') as title, coalesce(a.full_vendor_name, '') as full_vendor_name, coalesce(a.availability, '') as availability, coalesce(fund.name, '') fund"
-		projection = "a.id, a.author, a.title, a.full_vendor_name, a.availability, o.decision, fund.name||':'||fund.title fund"
+		// projection = "a.id, coalesce(a.author, '') as author, coalesce(a.title, '') as title, coalesce(a.full_vendor_name, '') as full_vendor_name, coalesce(a.availability, '') as availability, a.holdings_count, coalesce(fund.name, '') fund"
+		projection = "a.id, a.author, a.title, a.full_vendor_name, a.availability, a.holdings_count, o.decision, fund.name||':'||fund.title fund"
 	case "count(*)":
 		projection = "count(*)"
 	}
@@ -406,7 +406,7 @@ func evalExprOptAttr(db *dbx.DB, a, b *strings.Builder, expr Node, state evalSta
 func attrSQL(a, b *strings.Builder, attr string) {
 	a.WriteString(attr)
 	switch attr {
-	case "id", "author", "title", "full_vendor_name", "availability":
+	case "id", "author", "title", "full_vendor_name", "availability", "holdings_count":
 		b.WriteRune('a')
 		b.WriteRune('.')
 		b.WriteString(attr)